#12 feat: Refactor Benefits section to use GenericCarousel for improved performance and maintainability
