+++ frontend/src/components/layout/BenefitsSection.tsx
@@ -2,8 +2,8 @@ import BenefitsCarousel from "../sections/BenefitsCarousel";
 
 export default function BenefitsSection() {
   return (
-    <section id="benefits" className="w-full min-h-[440px] sm:min-h-[540px] md:min-h-[640px] lg:min-h-[821px] bg-purple-100 flex flex-col items-center py-8 sm:py-12 md:py-16 px-4">
-      <div className="flex justify-center mb-12 sm:mb-8 px-4">
+    <section id="benefits" className="w-full min-h-[440px] sm:min-h-[540px] md:min-h-[640px] lg:min-h-[600px] bg-purple-100 flex flex-col items-center px-4 sm:px-8 md:px-16 lg:px-24 xl:px-32 py-8 sm:py-12 md:py-16">
+      <div className="flex justify-center mb-12 sm:mb-8">
         <h2 className="text-purple-900 text-2xl sm:text-3xl md:text-4xl lg:text-5xl font-semibold font-['Montserrat'] leading-tight text-center">
           Por que escolher o <span className="text-purple-600">Pole Dance</span>?
         </h2>
