+++ frontend/src/data/benefitsData.tsx
@@ -1,0 +1,41 @@
+import { 
+  StrengthIcon, 
+  FlexibilityIcon, 
+  ConfidenceIcon, 
+  FunIcon, 
+  ExpressionIcon, 
+  CommunityIcon 
+} from "../components/sections/BenefitIcons";
+
+export const benefitsData = [
+  {
+    icon: <StrengthIcon />,
+    title: "Força",
+    description: "Desenvolva músculos e resistência enquanto se diverte em cada aula."
+  },
+  {
+    icon: <FlexibilityIcon />,
+    title: "Flexibilidade",
+    description: "Melhore sua mobilidade e descubra novas formas de movimentar o corpo."
+  },
+  {
+    icon: <ConfidenceIcon />,
+    title: "Confiança",
+    description: "Sinta-se mais segura e orgulhosa a cada conquista na barra."
+  },
+  {
+    icon: <FunIcon />,
+    title: "Diversão",
+    description: "Exercite-se de forma leve, criativa e sempre acompanhada de boa música."
+  },
+  {
+    icon: <ExpressionIcon />,
+    title: "Expressão",
+    description: "Explore sua sensualidade e arte corporal, do seu jeito."
+  },
+  {
+    icon: <CommunityIcon />,
+    title: "Comunidade",
+    description: "Faça parte de um grupo acolhedor que apoia e celebra suas vitórias."
+  }
+];
